+++ CHANGES.rst
@@ -1,6 +1,8 @@
 0.2.0 (WIP)
 -----------
 
+- Feature GH-1: Add support to TAP mode of OpenVPN servers by parsing virtual
+  addresses as MAC and IP both.
 - Fix GH-2: **BREAK-COMPATIBILITY** Use real or virtual addresses as the key
   of client_list and routing_table, instead of using common name.